+++ lib/features/appConfiguration/domain/usecase/get_last_app_terms_version_use_case.dart
@@ -1,0 +1,13 @@
+import 'package:zachranobed/features/appConfiguration/domain/repository/app_configuration_repository.dart';
+
+class GetLastAppTermsVersionUseCase {
+  final AppConfigurationRepository _appConfigurationRepository;
+
+  GetLastAppTermsVersionUseCase(this._appConfigurationRepository);
+
+  Future<int?> getLastAppTerms() async {
+    var data = await _appConfigurationRepository.fetchData();
+
+    return data?.lastAppTermsVersion;
+  }
+}