+++ lib/features/menu/domain/model/contact.freezed.dart
@@ -20,7 +20,9 @@ mixin _$Contact {
   String? get position => throw _privateConstructorUsedError;
   String? get phoneNumber => throw _privateConstructorUsedError;
 
-  @JsonKey(ignore: true)
+  /// Create a copy of Contact
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
   $ContactCopyWith<Contact> get copyWith => throw _privateConstructorUsedError;
 }
 
@@ -42,6 +44,8 @@ class _$ContactCopyWithImpl<$Res, $Val extends Contact>
   // ignore: unused_field
   final $Res Function($Val) _then;
 
+  /// Create a copy of Contact
+  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -84,6 +88,8 @@ class __$$$ContactImplCopyWithImpl<$Res>
       _$$ContactImpl _value, $Res Function(_$$ContactImpl) _then)
       : super(_value, _then);
 
+  /// Create a copy of Contact
+  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -141,7 +147,9 @@ class _$$ContactImpl implements $Contact {
   @override
   int get hashCode => Object.hash(runtimeType, name, position, phoneNumber);
 
-  @JsonKey(ignore: true)
+  /// Create a copy of Contact
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
   @override
   @pragma('vm:prefer-inline')
   _$$$ContactImplCopyWith<_$$ContactImpl> get copyWith =>
@@ -160,8 +168,11 @@ abstract class $Contact implements Contact {
   String? get position;
   @override
   String? get phoneNumber;
+
+  /// Create a copy of Contact
+  /// with the given fields replaced by the non-null parameter values.
   @override
-  @JsonKey(ignore: true)
+  @JsonKey(includeFromJson: false, includeToJson: false)
   _$$$ContactImplCopyWith<_$$ContactImpl> get copyWith =>
       throw _privateConstructorUsedError;
 }