+++ lib/models/dto/appconfig_dto.g.dart
@@ -1,0 +1,16 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'appconfig_dto.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+AppConfigDto _$AppConfigDtoFromJson(Map<String, dynamic> json) => AppConfigDto(
+      minimumAppVersion: json['minimumAppVersion'] as String,
+    );
+
+Map<String, dynamic> _$AppConfigDtoToJson(AppConfigDto instance) =>
+    <String, dynamic>{
+      'minimumAppVersion': instance.minimumAppVersion,
+    };